Bugfix on Linux fs mount
The script worked only if it was run from `/` folder
because `mnt` folder were refered as using a relative path.

Adding `set -e` at the very beginning of the script
is a best practice to always use in bash scripting unless
there is a specific reason. It helps to avoid hidden errors.