Pass parameters through environment variable
Reading variables through /proc/cmdline is good for deployment automation,
especially when booting a live distribution from the network as an installation
environment.

However, we should be able to run the installer passing parameters directly
to the script because there no reason to enforce this way of doing. We should
be able to pass parameters even if it was not previoulsy configured in the
kernel parameter line.

It's also usefull for testing purpose during development because we can run
different tests without the need of rebooting the machine.