Use shfmt to format code in an autoimated way
shfmt implements the same rules as the one we follow on the project.

It also have a Visual Studio code extension to format code
automatically.

We can now check code formatting into the CI.

Also, we have to be careful with YAML template in the code. YAML does not
accept TAB indentation. So, when generating the file /etc/netplan/network.yaml
we have to use spaces to indent YAML into the heredoc content.