SHEEP_CI_RUNNER_IP is specific to any runner
It is actually specific to any group of runner. All runners running
on the same Linux have the same IP. We'll add runner in Paris in
addition to runners in Newark. So, they're running on a different
machine and so far we had no way to configure the SHEEP_CI_RUNNER_IP.

In the CI, the IP will now be read from a configuration file. No need
anymore to set this IP into GitLab CI descriptor